Basic working example
Here, I built a basic working example, where we use the C code to
calculate the next values for the animation.

Not totally right, because the state is still held in js. But in the
next commits, I want to try building a small simulation in C, that is
displayed in the web via three.js.